requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
psycopg2-binary>=2.9.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.0
gunicorn==21.2.0
python-dotenv==1.0.1 